requirements.txt
# Core Trading & Market Data
alpaca-py==0.21.0
alpaca-trade-api==3.1.1
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3

# Options Analytics
py_vollib==1.0.1
scipy==1.12.0
mibian==0.1.3

# Database & Storage
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Scheduling & Async
schedule==1.2.1
APScheduler==3.10.4
asyncio==3.4.3

# ML & Analytics
scikit-learn==1.4.0
shap==0.44.1
statsmodels==0.14.1

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0

# Utils
python-dateutil==2.8.2
pytz==2024.1


